+++ lib/flutter_advanced_switch.dart
@@ -1,7 +1,277 @@
 import 'package:flutter/animation.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
 
-part 'src/controller.dart';
+class AdvancedSwitch extends StatefulWidget {
+  const AdvancedSwitch({
+    Key? key,
+    this.controller,
+    this.activeColor = const Color(0xFF4CAF50),
+    this.inactiveColor = const Color(0xFF9E9E9E),
+    this.activeChild,
+    this.inactiveChild,
+    this.activeImage,
+    this.inactiveImage,
+    this.borderRadius = const BorderRadius.all(const Radius.circular(15)),
+    this.width = 50.0,
+    this.height = 30.0,
+    this.enabled = true,
+    this.disabledOpacity = 0.5,
+    this.thumb,
+  }) : super(key: key);
 
-part 'src/widget.dart';
+  /// Determines if widget is enabled
+  final bool enabled;
+
+  /// Determines current state.
+  final ValueNotifier<bool>? controller;
+
+  /// Determines background color for the active state.
+  final Color activeColor;
+
+  /// Determines background color for the inactive state.
+  final Color inactiveColor;
+
+  /// Determines label for the active state.
+  final Widget? activeChild;
+
+  /// Determines label for the inactive state.
+  final Widget? inactiveChild;
+
+  /// Determines background image for the active state.
+  final ImageProvider? activeImage;
+
+  /// Determines background image for the inactive state.
+  final ImageProvider? inactiveImage;
+
+  /// Determines border radius.
+  final BorderRadius borderRadius;
+
+  /// Determines width.
+  final double width;
+
+  /// Determines height.
+  final double height;
+
+  /// Determines opacity of disabled control.
+  final double disabledOpacity;
+
+  /// Thumb widget.
+  final Widget? thumb;
+
+  @override
+  _AdvancedSwitchState createState() => _AdvancedSwitchState();
+}
+
+class _AdvancedSwitchState extends State<AdvancedSwitch>
+    with SingleTickerProviderStateMixin {
+  static const _duration = Duration(milliseconds: 250);
+  late ValueNotifier<bool> _controller;
+  late AnimationController _animationController;
+  late Animation<Offset> _slideAnimation;
+  late Animation<Color?> _colorAnimation;
+  late double _thumbSize;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _controller = widget.controller ?? ValueNotifier<bool>(false);
+
+    _controller.addListener(() {
+      if (_controller.value) {
+        _animationController.forward();
+      } else {
+        _animationController.reverse();
+      }
+    });
+
+    _animationController = AnimationController(
+      vsync: this,
+      duration: _duration,
+      value: _controller.value ? 1.0 : 0.0,
+    );
+
+    _initAnimation();
+  }
+
+  @override
+  void didUpdateWidget(covariant AdvancedSwitch oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    _initAnimation();
+  }
+
+  void _initAnimation() {
+    _thumbSize = widget.height;
+    final offset = widget.width / 2 - _thumbSize / 2;
+
+    final animation = CurvedAnimation(
+      parent: _animationController,
+      curve: Curves.easeInOut,
+    );
+
+    _slideAnimation = Tween<Offset>(
+      begin: Offset(-offset, 0),
+      end: Offset(offset, 0),
+    ).animate(animation);
+
+    _colorAnimation = ColorTween(
+      begin: widget.inactiveColor,
+      end: widget.activeColor,
+    ).animate(animation);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final labelSize = widget.width - _thumbSize;
+    final containerSize = labelSize * 2 + _thumbSize;
+
+    return MouseRegion(
+      cursor: SystemMouseCursors.click,
+      child: GestureDetector(
+        onTap: _handlePressed,
+        child: Opacity(
+          opacity: widget.enabled ? 1 : widget.disabledOpacity,
+          child: AnimatedBuilder(
+            animation: _animationController,
+            builder: (_, child) {
+              return Container(
+                width: widget.width,
+                height: widget.height,
+                clipBehavior: Clip.antiAlias,
+                decoration: BoxDecoration(
+                  borderRadius: widget.borderRadius,
+                  color: _colorAnimation.value,
+                ),
+                child: child,
+              );
+            },
+            child: Stack(
+              children: [
+                if (widget.activeImage != null || widget.inactiveImage != null)
+                  ValueListenableBuilder<bool>(
+                    valueListenable: _controller,
+                    builder: (_, __, ___) {
+                      return AnimatedCrossFade(
+                        crossFadeState: _controller.value
+                            ? CrossFadeState.showSecond
+                            : CrossFadeState.showFirst,
+                        duration: _duration,
+                        firstChild: Image(
+                          width: widget.width,
+                          height: widget.height,
+                          image: widget.inactiveImage ?? widget.activeImage!,
+                          fit: BoxFit.cover,
+                        ),
+                        secondChild: Image(
+                          width: widget.width,
+                          height: widget.height,
+                          image: widget.activeImage ?? widget.inactiveImage!,
+                          fit: BoxFit.cover,
+                        ),
+                      );
+                    },
+                  ),
+                AnimatedBuilder(
+                  animation: _animationController,
+                  builder: (context, child) {
+                    return Transform.translate(
+                      offset: _slideAnimation.value,
+                      child: child,
+                    );
+                  },
+                  child: OverflowBox(
+                    minWidth: containerSize,
+                    maxWidth: containerSize,
+                    minHeight: widget.height,
+                    maxHeight: widget.height,
+                    child: Row(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        IconTheme(
+                          data: const IconThemeData(
+                            color: Color(0xFFFFFFFF),
+                            size: 20,
+                          ),
+                          child: DefaultTextStyle(
+                            style: const TextStyle(
+                              color: Color(0xFFFFFFFF),
+                              fontWeight: FontWeight.w500,
+                              fontSize: 12,
+                            ),
+                            child: Container(
+                              width: labelSize,
+                              height: widget.height,
+                              alignment: Alignment.center,
+                              child: widget.activeChild,
+                            ),
+                          ),
+                        ),
+                        Container(
+                          margin: const EdgeInsets.all(2),
+                          width: _thumbSize - 4,
+                          height: _thumbSize - 4,
+                          child: widget.thumb ??
+                              Container(
+                                decoration: BoxDecoration(
+                                  color: const Color(0xFFFFFFFF),
+                                  borderRadius: widget.borderRadius
+                                      .subtract(BorderRadius.circular(1)),
+                                  boxShadow: const [
+                                    BoxShadow(
+                                      color: Color(0x42000000),
+                                      blurRadius: 8,
+                                    ),
+                                  ],
+                                ),
+                              ),
+                        ),
+                        IconTheme(
+                          data: const IconThemeData(
+                            color: Color(0xFFFFFFFF),
+                            size: 20,
+                          ),
+                          child: DefaultTextStyle(
+                            style: const TextStyle(
+                              color: Color(0xFFFFFFFF),
+                              fontWeight: FontWeight.w500,
+                              fontSize: 12,
+                            ),
+                            child: Container(
+                              width: labelSize,
+                              height: widget.height,
+                              alignment: Alignment.center,
+                              child: widget.inactiveChild,
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  void _handlePressed() {
+    if (widget.controller != null && widget.enabled) {
+      _controller.value = !_controller.value;
+    }
+  }
+
+  @override
+  void dispose() {
+    _animationController.dispose();
+
+    if (widget.controller == null) {
+      _controller.dispose();
+    }
+
+    super.dispose();
+  }
+}